requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.5.30
charset-normalizer==2.0.3; python_version >= '3'
coreapi==2.3.3
coreschema==0.0.4
django==3.2.5
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2; python_version >= '3'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
packaging==21.0; python_version >= '3.6'
pyjwt==2.1.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ruamel.yaml.clib==0.2.6; python_version < '3.10' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.10; python_version >= '3'
sqlparse==0.4.1; python_version >= '3.5'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
